Move atom plotting function to horton.scripts.atomdb
